+++ packages/stepper/src/StepIcon/types.ts
@@ -1,0 +1,7 @@
+import { StepState } from '../types';
+
+export interface StepIconProps {
+  state: StepState;
+  size: number;
+  content?: string | number;
+}
