+++ packages/stepper/src/StepLabel/types.ts
@@ -1,0 +1,3 @@
+import { InternalStepProps } from '../InternalStep/types';
+
+export type StepLabelProps = Pick<InternalStepProps, 'state'>;
