+++ packages/stepper/src/Stepper/styles.ts
@@ -1,0 +1,13 @@
+import { css } from '@leafygreen-ui/emotion';
+
+const baseStyles = css`
+  list-style: none;
+  padding-inline-start: 0;
+  width: 100%;
+  display: flex;
+  & > * {
+    flex: 1;
+  }
+`;
+
+export default baseStyles;
