+++ packages/stepper/src/StepperContext/types.ts
@@ -1,0 +1,8 @@
+import React, { Dispatch } from 'react';
+
+export interface StepperContextValues {
+  darkMode: boolean;
+  setDarkMode: Dispatch<React.SetStateAction<boolean>>;
+  stepIconClassName: string;
+  stepLabelClassName: string;
+}
