+++ packages/stepper/src/index.ts
@@ -1,6 +1,6 @@
 import Stepper from './Stepper';
+import type { StepperProps } from './Stepper/types';
 import Step from './Step';
-import type { StepperProps } from './types';
 export { Step };
 export type { StepperProps };
 export default Stepper;
